Reorganize well-known traits tests
