--- tests/test/builtin_impls.rs
@@ -1,5 +1,0 @@
-mod clone;
-mod copy;
-mod fn_family;
-mod sized;
-mod unsize;
