+++ tests/test/fn_def.rs
@@ -166,3 +166,62 @@ fn generic_fn_implements_fn_traits() {
         }
     }
 }
+
+#[test]
+fn fn_defs() {
+    test! {
+        program {
+            trait Foo { }
+
+            struct Bar { }
+
+            struct Xyzzy { }
+            impl Foo for Xyzzy { }
+
+            fn baz<T>(quux: T) -> T
+                where T: Foo;
+
+            fn garply(thud: i32) -> i32;
+        }
+
+        goal {
+            WellFormed(baz<Bar>)
+        } yields {
+            "No possible solution"
+        }
+
+        goal {
+            WellFormed(baz<Xyzzy>)
+        } yields {
+            "Unique"
+        }
+
+        goal {
+            WellFormed(garply)
+        } yields {
+            "Unique"
+        }
+
+    }
+}
+
+#[test]
+fn fn_def_implied_bounds_from_env() {
+    test! {
+        program {
+            trait Foo { }
+
+            struct Bar { }
+            impl Foo for Bar { }
+
+            fn baz<T>() where T: Foo;
+        }
+        goal {
+            if (FromEnv(baz<Bar>)) {
+                Bar: Foo
+            }
+        } yields {
+            "Unique"
+        }
+    }
+}
